📦️chore: types 이름 변경
