+++ src/app/types/commit.ts
@@ -1,6 +1,15 @@
-type CommitInfo = {
-  id: string;
-  message: string;
-};
+interface CommitFile {
+  filename: string;
+  status: string;
+  patch: string | null;
+}
 
-export type { CommitInfo };
+interface CommitDetail {
+  sha: string;
+  author: string;
+  date: string;
+  message: string;
+  files: CommitFile[] | null;
+}
+
+export type { CommitFile, CommitDetail };
