--- src/app/types/githubCommit.ts
@@ -1,15 +1,0 @@
-interface CommitFile {
-  filename: string;
-  status: string;
-  patch: string | null;
-}
-
-interface CommitDetail {
-  sha: string;
-  author: string;
-  date: string;
-  message: string;
-  files: CommitFile[] | null;
-}
-
-export type { CommitFile, CommitDetail };
